requirements.txt
PyQt5
requests
packaging
fuzzywuzzy
beautifulsoup4
PyPDF2
pytest
python-dotenv